Refactor #13 : 역할과 책임을 생각하여 GameProcessor -> BallFactory로 인터페이스 이름 변경
