+++ src/main/java/baseball/domain/ball/BallFactory.java
@@ -1,0 +1,7 @@
+package baseball.domain.ball;
+
+import java.util.List;
+
+public interface BallFactory {
+    List<Integer> numbersToIntegerList(String numbers);
+}
