--- src/main/java/baseball/domain/game/GameProcessor.java
@@ -1,7 +1,0 @@
-package baseball.domain.game;
-
-import java.util.List;
-
-public interface GameProcessor {
-    List<Integer> numbersToList(String numbers);
-}
